requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
oandapyV20==0.6.3
pysimplemodel==2.3.3
python-dateutil==2.8.1
python-status==1.0.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
